chore: Refactor the send email function and controller
