+++ server/src/controllers/email/index.js
@@ -1,48 +1,18 @@
-const nodemailer = require('nodemailer')
-const { google } = require('googleapis')
-const OAuth2 = google.auth.OAuth2
-
-const oauth2Client = new OAuth2(
-  process.env.CLIENT_ID,
-  process.env.CLIENT_SECRET,
-  process.env.REDIRECT_URI
-)
-
-oauth2Client.setCredentials({ refresh_token: process.env.REFRESH_TOKEN })
+const _sendemail = require('../../utils/email')
 
 // Send email using gmail OAuth
 module.exports.sendEmail = async (req, res, next) => {
+  const { to, from, subject, text } = req.body
+
+  // TO-DO: Use a detailed server-side validation later using joi
+  if (to === undefined || from === undefined || subject === undefined || text === undefined) {
+    return res.status(400).send('Missing parameter/s')
+  }
+
   try {
-    const { token } = await oauth2Client.getAccessToken()
-    const { to, from, subject, text } = req.body
-
-    const newData = {
-      to,
-      subject,
-      text,
-      from: from || process.env.EMAIL
-    }
-
-    const transporter = nodemailer.createTransport({
-      service: 'gmail',
-      auth: {
-        type: 'OAuth2',
-        user: process.env.CLIENT_USER,
-        clientId: process.env.CLIENT_ID,
-        clientSecret: process.env.CLIENT_SECRET,
-        refreshToken: process.env.REFRESH_TOKEN,
-        accessToken: token
-      }
-    })
-
-    transporter.sendMail(newData, (err, data) => {
-      if (err) {
-        return res.status(500).send(err.message)
-      } else {
-        return res.status(200).send('Email sent!')
-      }
-    })
+    const response = await _sendemail()
+    return res.status(200).send(response)
   } catch (err) {
-    return res.status(500).send(err.message)
+    return next(err.message)
   }
 }
