+++ server/src/controllers/index.js
@@ -16,6 +16,21 @@ const Email = require('./email')
  * @apiBody {String} from Sender's email address
  * @apiBody {String} subject Email subject/title
  * @apiBody {String} text Email text content
+ *
+ * @apiSuccess {String} message Message log of successful email sent.
+ *
+ * @apiExample {js} Example usage:
+ * const obj = {
+ *   data: {
+ *     to: 'you@gmail.com',
+ *     from: 'me@gmail.com',
+ *     subject: 'Hello!',
+ *     text: 'Hi, this is a test message.'
+ *   },
+ *   headers: {
+ *     Authorization: 'Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IjZhNGY4N2Z....'
+ *   }
+ * }
  */
 router.post('/email', Email.sendEmail)
